pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "address_bundler"
version = "0.1.0"
description = "Sort addresses into geographically grouped bundles"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Matt King", email = "kyrrigle@gmail.com" }
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "docopt",
  "icecream",
  "python-dotenv",
  "scikit-learn",
  "numpy",
  "geopy",
  "pdfplumber",
  "fuzzywuzzy",
  "prettytable",
  "reportlab",
  "peewee",
  "PyYAML",
  "tqdm",
  "py-staticmaps[cairo]",
  "Pillow<10.0",
  "usaddress",
  "scipy (>=1.11)",
  "fuzzywuzzy[speedup]>=0.18.0",
  "face_recognition",
  "face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models.git",
  "setuptools<81",
]

[tool.poetry.group.dev.dependencies]
black = "*"
pytest = "*"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]

[project.scripts]
ab-bundle = "address_bundler.main:main"
ab-signs = "lawn_signs.main:main"
ab-project = "common.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''